requirements.txt
langchain==0.1.14
langchain-community==0.0.30
langchain-google-genai==0.0.5
python-dotenv
google-cloud-bigquery
google-auth
google-generativeai
